style: workspace fmt + clippy sweep (1321 -> 83 non-policy warnings)
`cargo fmt --all --check` was failing on 18 files; it passes now. That is
CI's strict gate. The clippy gate (`[workspace.lints]`: `unsafe_code` and
`unused_must_use` deny) was already clean and stays clean -- this clears
warnings, not errors.

Of the 1321 warnings, 658 are `unwrap_used` / `panic`, which the workspace
sets to `warn` deliberately and which are overwhelmingly test code. Those
are left alone. The rest went 600 -> 83.

Mechanical, applied via `cargo clippy --fix` / `cargo fix` and verified by a
full workspace run: ~93 nested `if let` collapsed to let-chains (Rust 2024,
already used here), 23 `field_reassign_with_default` rewritten as struct
literals, plus `cloned_ref_to_slice_refs`, `manual_clamp`, `map_identity`,
`unnecessary_get_then_check` and the resulting `unused_mut`.

The doc lints were worth more than their count suggests -- 225 of them, and
clearing them surfaced three real defects rather than style nits:

- `Editor`'s 135-line doc block sat at the top of `editor.rs` with no item
  after it, so all of it rendered as `OverlayWake`'s documentation while
  `Editor` itself had none. Moved onto the struct.
- `diffthis_group`'s doc was attached to `scrollbind_group_id`, leaving
  `diffthis_group` undocumented.
- An orphaned block from a retired body was attaching to
  `do_completion_accept`.

27 further orphaned `///` blocks -- left behind when their item was
relocated, and silently attaching to whatever item followed -- were demoted
to `//`, keeping the history text without the mis-attachment. Over-indented
list continuations were also rendering as CODE BLOCKS in rustdoc rather
than prose.

Two escape fixes are behavioural-adjacent: `"...\03 days ago"` reads as an
octal escape (Rust has none -- it is NUL then `3`), now written `\x00`.

One clippy suggestion was WRONG and is reverted with a comment: `ptr_arg`
on `validate_color` wanted `&str`, but `#[validate(...)]` takes a
`ValidateFn<T>` where `T` is the option's declared type (`String`), and
`&str` does not coerce to that fn pointer.

Left as warnings, deliberately: 23 `type_complexity`, 13
`too_many_arguments`, 9 `result_large_err`, and a few peers. Each needs a
real refactor of dispatch/hot-path code rather than a rewrite, and CI's own
comment argues that non-blocking warnings are the right treatment for this
class. `result_large_err` on `editor_actor.rs` in particular is a genuine
signal worth keeping visible rather than silencing with 57 `#[allow]`s.

+++ crates/lattice-ui-gpui/src/window.rs
@@ -129,8 +129,8 @@ pub(crate) fn popup_outer_dims_px(viewport_w_px: f32, viewport_h_px: f32) -> (f3
 /// top+bottom + the bold/larger title row + .pb_2 header gap).
 /// Subtract from the popup's outer height to get the inner body area.
 ///
-/// 2026-06: the `───` separator row was removed (the bold, larger title
-/// + the `.pb_2()` gap separate the header from the body); the title row
+/// 2026-06: the `───` separator row was removed (the bold, larger title +
+/// the `.pb_2()` gap separate the header from the body); the title row
 /// is now `row_px * POPUP_TITLE_SCALE` tall (the larger font's line
 /// height), locked in the paint so this stays exact.
 pub(crate) fn popup_chrome_v_px(rem: f32, row_px: f32) -> f32 {
@@ -881,6 +881,7 @@ fn preview_char_color_rgb(
 ///   chord highlight)
 /// - source      → mauve (purple/magenta)
 /// - custom      → blue (fallback for plugin annotations)
+///
 /// T.6: resolve a completion-annotation color. The *base*
 /// (unselected) color now reads from the registered
 /// `completion.annotation.{kind,doc,keybinding,source,custom}`
@@ -3857,7 +3858,7 @@ impl Render for EditorView {
                 .is_some_and(|b| !b.is_empty());
             match (comp.docs_buffer_id, doc_popup) {
                 (Some(docs_id), Some(doc_popup)) if body_present => {
-                    let scroll = doc_popup.scroll as u32;
+                    let scroll = doc_popup.scroll;
                     let content_snap = rs
                         .buffers
                         .registry
@@ -5228,7 +5229,11 @@ fn transient_rows_gpui(
                             _ => None,
                         })
                         .unwrap_or(false);
-                    if v { "[x]".to_string() } else { "[ ]".to_string() }
+                    if v {
+                        "[x]".to_string()
+                    } else {
+                        "[ ]".to_string()
+                    }
                 }
                 // MG.43g: the TUI peer's variable rendering, in
                 // lockstep. Three states — `…` (not read yet) is not
